+++ lib/pbio/drv/counter/counter_lpf2.c
@@ -113,7 +113,7 @@ static const pbdrv_counter_funcs_t pbdrv_counter_lpf2_funcs = {
 };
 
 void pbdrv_counter_lpf2_init(pbdrv_counter_dev_t *devs) {
-    for (int i = 0; i < PBIO_ARRAY_SIZE(private_data); i++) {
+    for (size_t i = 0; i < PBIO_ARRAY_SIZE(private_data); i++) {
         const pbdrv_counter_lpf2_platform_data_t *pdata =
             &pbdrv_counter_lpf2_platform_data[i];
         private_data_t *priv = &private_data[i];
